refactor - ViewModel for each feature screen
