+++ composeApp/src/commonMain/kotlin/ui/screen/onboarding/MoodRateViewModel.kt
@@ -1,21 +1,13 @@
-package ui.viewmodel
+package ui.screen.onboarding
 
 import kotlinx.coroutines.launch
 import moe.tlaster.precompose.viewmodel.ViewModel
 import moe.tlaster.precompose.viewmodel.viewModelScope
 import repo.KeyValueStorageRepository
 
-class MainViewModel(
+class MoodRateViewModel(
     private val kvsRepo: KeyValueStorageRepository
 ) : ViewModel() {
-
-    fun saveFinishedOnboarding() {
-        kvsRepo.saveOnboardingFinished()
-    }
-
-    fun onboardingFinished(): Boolean {
-        return kvsRepo.onboardingFinished()
-    }
 
     fun setMoodStateOnboardingSeen() {
         viewModelScope.launch {
