+++ composeApp/src/commonMain/kotlin/ui/screen/reward/CongratulationViewModel.kt
@@ -1,0 +1,13 @@
+package ui.screen.reward
+
+import moe.tlaster.precompose.viewmodel.ViewModel
+import repo.KeyValueStorageRepository
+
+class CongratulationViewModel(
+    private val kvsRepo: KeyValueStorageRepository
+) : ViewModel() {
+
+    fun saveFinishedOnboarding() {
+        kvsRepo.saveOnboardingFinished()
+    }
+}
