+++ composeApp/src/commonMain/kotlin/ui/screen/splash/SplashViewModel.kt
@@ -1,0 +1,13 @@
+package ui.screen.splash
+
+import moe.tlaster.precompose.viewmodel.ViewModel
+import repo.KeyValueStorageRepository
+
+class SplashViewModel(
+    private val kvsRepo: KeyValueStorageRepository
+) : ViewModel() {
+
+    fun onboardingFinished(): Boolean {
+        return kvsRepo.onboardingFinished()
+    }
+}
